Cleaned up file structure
